# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Backend API
  description: Production-grade invoice/billing platform API
  version: 1.0.0
  contact:
    name: API Support
    email: support@invoiceapp.local
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api/v1
    description: API v1 base path
...
paths:
...
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset code sent if email exists
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with code
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification code
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification code sent
...
  /api/v1/auth/profile:
    put:
      tags: [Auth]
      summary: Update profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
...
  /api/v1/business-profiles/{id}/logo:
    post:
      tags: [Business]
      summary: Get logo upload URL
      operationId: getLogoUploadUrl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [email, confirmation_code, new_password]
      properties:
        email:
          type: string
          format: email
        confirmation_code:
          type: string
        new_password:
          type: string
          minLength: 8
    VerifyEmailRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
    UpdateProfileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
          minLength: 8
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, accountant, viewer]
        business_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UploadUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
...
security:
  - bearerAuth: []